Extracted Class from DurationEditor
